add check to ensure that we do not reprocess twice

+++ app/workers/reprocess_paper.rb
@@ -1,9 +1,10 @@
 class ReprocessPaper
   include Sidekiq::Worker
 
-  def perform(doi)
+  def perform(doi, days)
     paper = PaperResult.find_by(doi: doi)
-    return if paper.nil?
+    # check to make sure the paper still exists AND it hasn't been updated in the meantime
+    return if paper.nil? || paper.updated_at > days.days.ago
     paper.destroy!
     AnalyzePaper.perform_async(doi)
   end
